파일명, 변수 및 기타 요소 명칭 변경
