+++ src/pages/index.ts
@@ -1,9 +1,17 @@
 import Main from "./Main";
 import Login from "./Login";
 import Register from "./Register";
-import PasswordSearch from "./PasswordSearch";
-import PasswordChange from "./PasswordChange.tsx";
+import FindPassword from "./FindPassword";
+import ChangePassword from "./ChangePassword";
 import Template from "./Template";
 import Community from "./Community";
 
-export { Main, Login, Register, PasswordSearch, PasswordChange, Template, Community };
+export {
+  Main,
+  Login,
+  Register,
+  FindPassword,
+  ChangePassword,
+  Template,
+  Community,
+};
